+++ plugins/coditary/conan/1.0.0/src/defaults.lua
@@ -78,7 +78,7 @@ end
 
 M.build_profiles = {
     code = {
-        scope = "code",
+        scope = "app",
         build_type_env = true,
         cmake_preset_from_build_type = true,
         build_tree_from_build_type = true,